Extract messages into context provider
