+++ src/Room/Messages.tsx
@@ -1,11 +1,8 @@
-import React, { createRef, useEffect, useState } from 'react'
-import { useQuery } from '@apollo/react-hooks'
-import moment from 'moment'
+import React, { createRef, useEffect } from 'react'
 import { Box } from 'rebass'
 
-import { useWebsocketContext } from 'Context'
-
-import { MESSAGES_QUERY, MessagesQuery, Message as MessageType } from './graphql'
+import { useMessagesContext } from './MessagesContextProvider'
+import { Message as MessageType } from './graphql'
 import Message from './Message'
 
 type MessageGroup = [string, MessageType[]]
@@ -89,35 +86,9 @@ const MessageGroup: React.FC<{ messageGroup: MessageGroup }> = ({ messageGroup }
   )
 }
 
-const messagesFrom = moment()
-  .subtract(2, 'days')
-  .toISOString()
 const Messages: React.FC = () => {
-  const { data, loading } = useQuery<MessagesQuery['data'], MessagesQuery['vars']>(MESSAGES_QUERY, {
-    variables: { from: messagesFrom },
-    fetchPolicy: 'network-only',
-  })
-  const [messages, setMessages] = useState<MessageType[]>([])
-  const [newMessage, setNewMessage] = useState<MessageType | null>(null)
+  const { messages } = useMessagesContext()
   const chat = createRef<HTMLDivElement>()
-
-  useEffect(() => {
-    if (!data) {
-      return
-    }
-
-    setMessages(data.messages)
-  }, [data])
-
-  useEffect(() => {
-    if (!newMessage) {
-      return
-    }
-
-    const newMessages = [...messages, newMessage].sort((prev, next) => (prev.createdAt > next.createdAt ? 1 : 0))
-    setMessages(newMessages)
-    setNewMessage(null)
-  }, [messages, newMessage])
 
   useEffect(() => {
     if (!chat || !chat.current) {
@@ -126,17 +97,6 @@ const Messages: React.FC = () => {
 
     chat.current.scrollTop = chat.current.scrollHeight
   }, [chat, messages])
-
-  const websocket = useWebsocketContext()
-  useEffect(() => {
-    return websocket.subscribeToMessage(message => {
-      setNewMessage(message)
-    })
-  }, [websocket])
-
-  if (loading) {
-    return <p>Loading...</p>
-  }
 
   const groupedMessages = groupMessagesByRecord(messages)
