+++ src/Room/Room.tsx
@@ -14,6 +14,7 @@ import Main from './Main'
 import { ROOM_ACTIVATE, RoomActivate } from './graphql'
 import ApprovalContextProvider from './ApprovalContextProvider'
 import CurrentRecordContextProvider from './CurrentRecordContextProvider'
+import MessagesContextProvider from './MessagesContextProvider'
 import PlaylistRecordContextProvider from './PlaylistRecordContextProvider'
 
 const Room: React.FC = () => {
@@ -42,24 +43,26 @@ const Room: React.FC = () => {
     <ApprovalContextProvider>
       <PlaylistRecordContextProvider>
         <CurrentRecordContextProvider>
-          <VolumeContextProvider>
-            <Flex
-              sx={{
-                alignItems: 'top',
-                bg: 'background',
-                flexDirection: ['column', 'row'],
-                minHeight: '100vh',
-                mx: 'auto',
-                position: 'relative',
-              }}
-            >
-              <SideNav>
-                <Keyboard />
-              </SideNav>
-              <Main room={data.roomActivate.room} />
-              <Chat />
-            </Flex>
-          </VolumeContextProvider>
+          <MessagesContextProvider>
+            <VolumeContextProvider>
+              <Flex
+                sx={{
+                  alignItems: 'top',
+                  bg: 'background',
+                  flexDirection: ['column', 'row'],
+                  minHeight: '100vh',
+                  mx: 'auto',
+                  position: 'relative',
+                }}
+              >
+                <SideNav>
+                  <Keyboard />
+                </SideNav>
+                <Main room={data.roomActivate.room} />
+                <Chat />
+              </Flex>
+            </VolumeContextProvider>
+          </MessagesContextProvider>
         </CurrentRecordContextProvider>
       </PlaylistRecordContextProvider>
     </ApprovalContextProvider>
